+++ gremlin-server/src/test/java/org/apache/tinkerpop/gremlin/process/remote/RemoteGraphProcessComputerTest.java
@@ -22,7 +22,7 @@ package org.apache.tinkerpop.gremlin.process.remote;
 import org.apache.tinkerpop.gremlin.GraphProviderClass;
 import org.apache.tinkerpop.gremlin.driver.remote.RemoteGraphComputerProvider;
 import org.apache.tinkerpop.gremlin.process.ProcessComputerSuite;
-import org.apache.tinkerpop.gremlin.process.remote.RemoteGraph;
+import org.apache.tinkerpop.gremlin.structure.RemoteGraph;
 import org.junit.runner.RunWith;
 
 /**
